Update picture.md logic for responsive images
I have read and reread MDN on this topic re responsive images, and I think the logic of the comments etc in the original is reversed, so I propose this change. 

Setting `media="(min-width: 600px)"` seems to mean that the image will be used for screens over 600px, hence it should be used for the large images section. 

It appears it's possible to use `max-width` also, but it seems from reading MDN that it's correct to list the larger images first in the final picture element. 

Finally, this is an example of "art direction" as you showed me before @oscarotero , Lume will use the original file three file names when generating, I think, appending `-1000w` etc to the original filename. I edited the filenames to include `-big`, `-detail` and `-mini`.