Add basic usage example for `lume_markdown_plugins`
As a new user of both Deno and Lume, I found the doc missing some
basic examples. This commit adds one for lume_markdown_plugins.

markdown-it-footnote was written in JavaScript and has not been
actively maintained (last commited 2 years ago). This is a valid
chance (I think) to recomment plugins written in TypesScript and
especially those that are designed to be used with Lume.

I also found @mdit/plugin-footnote to be great to use with Lume and
is a good drop-in replacement for markdown-it-footnote. We may use
it in the section of Hooks as an exmple.